+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1124](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1124)
 - Added tooltip for line graph
   [GEOD-1126](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1126)
+- Added graph options in the detail page for starting graph at 0 or changing the x extent.
+  [GEOD-1120](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1120)
 
 ### Changed
 - Changed the Explore Page Pills to be larger and have more descriptive text